# swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Book Marketplace API"
  description: "API for managing books in the Book Marketplace application"
basePath: "/api"
...
definitions:
  Book:
    type: "object"
    properties:
      bookId:
        type: "integer"
      ISBN:
        type: "string"
      title:
        type: "string"
      author:
        type: "string"
      edition:
        type: "integer"
      publicationYear:
        type: "integer"
      language:
        type: "string"
      publisher:
        type: "string"
      originalPrice:
        type: "number"
      currentPrice:
        type: "number"
      description:
        type: "string"
      category:
        type: "string"
      conditionDescription:
        type: "string"
      condition:
        type: "string"
      status:
        type: "string"
...
  ErrorResponse:
    type: "object"
    properties:
      message:
        type: "string"
paths:
  /books/:
    get:
      tags:
        - "Books"
      summary: "Retrieve all books"
      description: "Retrieves a list of all books available in the marketplace."
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        204:
          description: "No books found"
        500:
          description: "Internal server error"
  /books/getAvailableBooks:
    get:
      tags:
        - "Books"
      summary: "Retrieve available books"
      description: "Retrieves a list of available books in the marketplace."
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        204:
          description: "No available books found"
        500:
          description: "Internal server error"
  /books/getBookById:
    get:
      tags:
        - "Books"
      summary: "Retrieve a book by ID"
      description: "Retrieves a book by its ID."
      parameters:
        - name: "bookId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/getBooksByCategory:
    get:
      tags:
        - "Books"
      summary: "Retrieve books by category"
      description: "Retrieves a list of books in the specified category."
      parameters:
        - name: "category"
          in: "query"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        204:
          description: "No books found in the specified category"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/getBookByISBN:
    get:
      tags:
        - "Books"
      summary: "Retrieve a book by ISBN"
      description: "Retrieves a book by its ISBN."
      parameters:
        - name: "isbn"
          in: "query"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/searchBooks:
    get:
      tags:
        - "Books"
      summary: "Search for books by keyword"
      description: "Searches for books based on a keyword in their title or author."
      parameters:
        - name: "keyword"
          in: "query"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        204:
          description: "No books found with the specified keyword"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/addBook:
    post:
      tags:
        - "Books"
      summary: "Add a new book"
      description: "Adds a new book to the marketplace."
      parameters:
        - name: "book"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: "Book added successfully"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/updateBook:
    put:
      tags:
        - "Books"
      summary: "Update a book"
      description: "Updates an existing book."
      parameters:
        - name: "bookId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
        - name: "book"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: "Book updated successfully"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/deleteBook:
    delete:
      tags:
        - "Books"
      summary: "Delete a book by ID"
      description: "Deletes a book by its ID."
      parameters:
        - name: "bookId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
      responses:
        200:
          description: "Book deleted successfully"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/buyBook:
    post:
      tags:
        - "Books"
      summary: "Purchase a book"
      description: "Allows a user to purchase a book by book ID."
      parameters:
        - name: "userId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
        - name: "bookId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
      responses:
        200:
          description: "Book purchased successfully"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/sellBook:
    post:
      tags:
        - "Books"
      summary: "Sell a book"
      description: "Allows a user to sell a book by book ID."
      parameters:
        - name: "userId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
        - name: "bookId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
      responses:
        200:
          description: "Book sold successfully"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
  /books/sellBookByISBN:
    post:
      tags:
        - "Books"
      summary: "Sell a book by ISBN"
      description: "Allows a user to sell a book by ISBN."
      parameters:
        - name: "userId"
          in: "query"
          required: true
          type: "integer"
          minimum: 1
        - name: "isbn"
          in: "query"
          required: true
          type: "string"
      responses:
        200:
          description: "Book sold successfully"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: "Internal server error"
...